requirements.txt
google-api-python-client==1.2
httplib2==0.8
wsgiref==0.1.2
Flask==0.10.1
Flask-Script==0.6.4
python-social-auth
flask-security
psycopg2
facebook-sdk
raven
gunicorn
flask-mongoengine==0.7.0
mongoengine==0.8.7
pymongo==2.6.3
flask-social
